+++ src/validation/depth.validator.ts
@@ -1,0 +1,7 @@
+import { InputValidator } from ".";
+
+export const depthValidator: InputValidator = (value: any) => {
+  if (!(typeof value === 'number' && value > 0)) {
+    throw new Error('Incorrect depth');
+  }
+};
