+++ src/validation/fo2.validator.ts
@@ -1,0 +1,7 @@
+import { InputValidator } from ".";
+
+export const fo2Validator: InputValidator = (value: any) => {
+  if (!(typeof value === 'number' && value > 0)) {
+    throw new Error('Incorrect fO2');
+  }
+};
